pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "boa"
version = "1.0.0"
description = "Bayesian Optimization Assistant - A server-based multi-objective optimization platform"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Ethan Schwartz", email = "ebuddy23@uw.edu"},
    {name = "Daniel Abdoue", email = "dabdoue@ucsd.edu"},
    {name = "Nicky Evans", email = "nickye17@mit.edu"},
    {name = "Tonio Buonassisi", email = "buonassi@mit.edu"},
]
maintainers = [
    {name = "Ethan Schwartz", email = "ebuddy23@uw.edu"},
]
keywords = [
    "bayesian-optimization",
    "multi-objective",
    "materials-science", 
    "thin-films",
    "machine-learning",
    "optimization",
    "gaussian-processes",
    "server",
    "api",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
    "Framework :: FastAPI",
]
requires-python = ">=3.10"
dependencies = [
    # Core scientific
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
    
    # Bayesian optimization
    "torch>=1.12.0",
    "gpytorch>=1.8.0",
    "botorch>=0.8.0",
    
    # Server & API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.12",
    
    # Database
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    
    # CLI & utilities
    "typer>=0.15.0",
    "rich>=13.9.0",
    "pyyaml>=6.0",
    "pydantic-settings>=2.0.0",
    
    # Visualization
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    
    # Legacy compatibility
    "pyDOE>=0.3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0",
    "ruff>=0.8.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
]
jupyter = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "notebook>=7.0.0",
]
shap = [
    "shap>=0.41.0",
]
all = [
    "boa[dev,jupyter,shap]",
]

[project.urls]
Homepage = "https://github.com/PV-Lab/BOA"
Repository = "https://github.com/PV-Lab/BOA"
Documentation = "https://github.com/PV-Lab/BOA#readme"
"Bug Tracker" = "https://github.com/PV-Lab/BOA/issues"

[project.scripts]
boa = "boa.cli.main:run"
# Legacy aliases
mobo-kit = "mobo_kit.cli:main"
mobo-kit-run = "mobo_kit.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=src/boa --cov-report=term-missing"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["src/boa"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
